tools/ontomanager/requirements.txt
requests~=2.31.0
molgenis-emx2-pyclient~=8.213.13
python-dotenv==1.0.0
pandas~=2.0.3
numpy~=1.25.1
tqdm~=4.66.1